Externalize Keystone secret key cache TTL
The secret key TTL is hard coded to 300 seconds (5 minutes).
For some use cases, the TTL could be increased, and as a consequence the number of requests to Keystone are reduced.

Therefore, we propose to externalize this configuration to an option to provide this flexibility to operators.
The default is maintained as 300 seconds (5 minutes).